fix(habit-task): unify endpoints under HabitTask controller
Merge POST /api/habit-tasks and evidence GET endpoints into a single HabitTask controller
to avoid duplicate HabitTask/HabitTasks groups in OpenAPI.
